reorganize seeder mod
